PDO Day
Used PDO in php to add functionaltiy where a players image will take you
to their page. Okay so that's not true but the functionality is there
and ready to go. I just need to start querying for individual players
and filling up the database. See the the blog for 12/12/17 for more
detailed info.